+++ apps/admin-x-settings/src/components/Settings.tsx
@@ -3,6 +3,7 @@ import React from 'react';
 import AdvancedSettings from './settings/advanced/AdvancedSettings';
 import EmailSettings from './settings/email/EmailSettings';
 import GeneralSettings from './settings/general/GeneralSettings';
+import GrowthSettings from './settings/growth/GrowthSettings';
 import MembershipSettings from './settings/membership/MembershipSettings';
 import SiteSettings from './settings/site/SiteSettings';
 
@@ -13,6 +14,7 @@ const Settings: React.FC = () => {
                 <GeneralSettings />
                 <SiteSettings />
                 <MembershipSettings />
+                <GrowthSettings />
                 <EmailSettings />
                 <AdvancedSettings />
             </div>
