+++ apps/admin-x-settings/src/components/providers/routing/modals.tsx
@@ -4,21 +4,21 @@ import type {RoutingModalProps} from '@tryghost/admin-x-framework/routing';
 import AboutModal from '../../settings/general/About';
 import AddIntegrationModal from '../../settings/advanced/integrations/AddIntegrationModal';
 import AddNewsletterModal from '../../settings/email/newsletters/AddNewsletterModal';
-import AddOfferModal from '../../settings/membership/offers/AddOfferModal';
-import AddRecommendationModal from '../../settings/membership/recommendations/AddRecommendationModal';
+import AddOfferModal from '../../settings/growth/offers/AddOfferModal';
+import AddRecommendationModal from '../../settings/growth/recommendations/AddRecommendationModal';
 import AmpModal from '../../settings/advanced/integrations/AmpModal';
 import AnnouncementBarModal from '../../settings/site/AnnouncementBarModal';
 import CustomIntegrationModal from '../../settings/advanced/integrations/CustomIntegrationModal';
 import DesignAndThemeModal from '../../settings/site/DesignAndThemeModal';
-import EditOfferModal from '../../settings/membership/offers/EditOfferModal';
-import EditRecommendationModal from '../../settings/membership/recommendations/EditRecommendationModal';
-import EmbedSignupFormModal from '../../settings/membership/embedSignup/EmbedSignupFormModal';
+import EditOfferModal from '../../settings/growth/offers/EditOfferModal';
+import EditRecommendationModal from '../../settings/growth/recommendations/EditRecommendationModal';
+import EmbedSignupFormModal from '../../settings/growth/embedSignup/EmbedSignupFormModal';
 import FirstpromoterModal from '../../settings/advanced/integrations/FirstPromoterModal';
 import HistoryModal from '../../settings/advanced/HistoryModal';
 import InviteUserModal from '../../settings/general/InviteUserModal';
 import NavigationModal from '../../settings/site/NavigationModal';
 import NewsletterDetailModal from '../../settings/email/newsletters/NewsletterDetailModal';
-import OffersModal from '../../settings/membership/offers/OffersModal';
+import OffersModal from '../../settings/growth/offers/OffersModal';
 import PinturaModal from '../../settings/advanced/integrations/PinturaModal';
 import PortalModal from '../../settings/membership/portal/PortalModal';
 import SlackModal from '../../settings/advanced/integrations/SlackModal';
