+++ apps/admin-x-settings/src/components/settings/growth/embedSignup/EmbedSignupSidebar.tsx
