+++ apps/admin-x-settings/src/components/settings/growth/recommendations/RecommendationIcon.tsx
